refactor refresh code and add refresh delete
