+++ src/main/scala/com/microsoft/hyperspace/actions/Constants.scala
@@ -30,10 +30,4 @@ object Constants {
   }
 
   val STABLE_STATES: Set[String] = Set(States.ACTIVE, States.DELETED, States.DOESNOTEXIST)
-
-  object RefreshModes {
-    val FULL = "FULL"
-    val QUICK = "QUICK"
-    val SMART = "SMART"
-  }
 }
